requirement.txt
absl-py==1.4.0
accelerate==0.20.3
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.7.2
astunparse==1.6.3
async-timeout==4.0.2
attrs==23.1.0
backports.zoneinfo==0.2.1
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
chex==0.1.6
click==8.1.3
cmake==3.26.4
colorama==0.4.6
commonmark==0.9.1
contourpy==1.1.0
cryptography==41.0.1
cycler==0.11.0
datasets==2.13.1
dill==0.3.6
Django==4.2.2
djangorestframework==3.14.0
dm-tree==0.1.8
etils==1.3.0
evaluate==0.4.0
filelock==3.12.2
flatbuffers==2.0.7
flax==0.5.3
fonttools==4.40.0
frozenlist==1.3.3
fsspec==2023.6.0
gast==0.4.0
google-auth==2.20.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.56.0
h5py==3.9.0
huggingface-hub==0.15.1
idna==3.4
importlib-metadata==6.7.0
importlib-resources==5.12.0
jax==0.3.6
jaxlib==0.3.5
Jinja2==3.1.2
joblib==1.2.0
keras==2.12.0
keras-nlp==0.5.2
kiwisolver==1.4.4
libclang==16.0.0
lit==16.0.6
Markdown==3.4.3
MarkupSafe==2.1.3
matplotlib==3.7.1
ml-dtypes==0.2.0
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.4
multiprocess==0.70.14
mypy-extensions==1.0.0
networkx==3.1
nltk==3.8.1
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
onnx==1.14.0
onnxconverter-common==1.13.0
opt-einsum==3.3.0
optax==0.1.4
packaging==23.1
pandas==2.0.2
Pillow==9.5.0
protobuf==4.23.3
psutil==5.9.5
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.15.1
pyparsing==3.1.0
pyre-extensions==0.0.29
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
regex==2023.6.3
requests==2.31.0
requests-oauthlib==1.3.1
responses==0.18.0
rich==11.2.0
rsa==4.9
safetensors==0.3.1
scikit-learn==1.2.0
scipy==1.10.1
sentence-transformers==2.2.2
sentencepiece==0.1.99
setfit==0.7.0
six==1.16.0
sqlparse==0.4.4
sympy==1.12
tensorboard==2.12.3
tensorboard-data-server==0.7.1
tensorflow==2.12.0
tensorflow-cpu==2.12.0
tensorflow-estimator==2.12.0
tensorflow-hub==0.13.0
tensorflow-io-gcs-filesystem==0.32.0
tensorflow-text==2.12.1
tensorstore==0.1.39
termcolor==2.3.0
tf2onnx==1.14.0
threadpoolctl==3.1.0
tokenizers==0.13.3
toolz==0.12.0
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
transformers==4.30.2
triton==2.0.0
typing-inspect==0.9.0
typing_extensions==4.6.3
tzdata==2023.3
urllib3==1.26.16
Werkzeug==2.3.6
wrapt==1.14.1
xformers==0.0.20
xxhash==3.2.0
yarl==1.9.2
zipp==3.15.0
